Rename src/types -> src/docs for clarity
